pyproject.toml
[project]
name = "arduino-dht11"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dearpygui>=2.2",
    "pyserial>=3.5",
]
